+++ cycletls/extensions.go
@@ -292,10 +292,12 @@ func IsGREASEValue(extID uint16) bool {
 }
 
 // CreateExtensionFromID creates an appropriate extension for the given ID
-func CreateExtensionFromID(extID uint16, tlsVersion uint16, components *JA4RComponents, disableGrease bool) utls.TLSExtension {
+func CreateExtensionFromID(extID uint16, tlsVersion uint16, components *JA4RComponents, disableGrease bool, serverName string) utls.TLSExtension {
 	switch extID {
 	case 0x0000: // Server Name Indication
-		return &utls.SNIExtension{}
+		return &utls.SNIExtension{
+			ServerName: serverName,
+		}
 	case 0x0005: // Status Request
 		return &utls.StatusRequestExtension{}
 	case 0x000a: // Supported Groups (Elliptic Curves)
@@ -342,7 +344,7 @@ func CreateExtensionFromID(extID uint16, tlsVersion uint16, components *JA4RComp
 	case 0x0010: // ALPN
 		alpnProtocols := []string{"h2", "http/1.1"}
 		if components != nil {
-			fmt.Printf("DEBUG: CreateExtensionFromID 0x0010 - components.ALPN = '%s'\n", components.ALPN)
+			
 			switch components.ALPN {
 			case "h2":
 				alpnProtocols = []string{"h2", "http/1.1"}
@@ -352,7 +354,7 @@ func CreateExtensionFromID(extID uint16, tlsVersion uint16, components *JA4RComp
 				alpnProtocols = []string{"h3", "h2", "http/1.1"}
 			}
 		}
-		fmt.Printf("DEBUG: CreateExtensionFromID 0x0010 - returning ALPN with protocols: %v\n", alpnProtocols)
+		
 		return &utls.ALPNExtension{
 			AlpnProtocols: alpnProtocols,
 		}